Update README.md model_fn first

+++ workshop_sections/mnist_series/cloudml/trainer/model.py
@@ -116,5 +116,5 @@ def evaluation(logits, labels):
   # It returns a bool tensor with shape [batch_size] that is true for
   # the examples where the label is in the top k (here k=1)
   # of all logits for that example.
-  correct = tf.nn.in_top_k(logits, labels, 1)
+  correct = tf.nn.in_top_k(logits, tf.cast(labels, tf.int32), 1)
   return tf.reduce_mean(tf.cast(correct, tf.float32))
